Add Django 2 and Python 3 versions
Add .travis.yml with a rather large matrix using tox-travis
testing only Django >= 1.8, and letting the pytest version
float as older pytest are not compatible with newer Django
or Python version.

Jobs for Django master are included but allowed to fail.

Adjust Django <=1.7 to keep the pin on pytest==3.0.2
however that will not work on many Python versions.